Feat: add alert to confirm remove of favorite dish

+++ components/FavoriteComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import Swipeout from 'react-native-swipeout';
@@ -27,15 +27,33 @@ class Favorites extends Component {
 
   render() {
     const { navigate } = this.props.navigation;
-   
+
     const renderMenuItem = ({ item, index }) => {
       const rightButton = [
         {
-          text: 'Delete', 
+          text: 'Delete',
           type: 'delete',
-          onPress: () => this.props.deleteFavorite(item.id)
+
+          onPress: () => {
+            Alert.alert(
+              'Delete Favorite?',
+              'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+              [
+                {
+                  text: 'Cancel',
+                  onPress: () => console.log(item.name + 'Not Deleted'),
+                  style: ' cancel'
+                },
+                {
+                  text: 'OK',
+                  onPress: () => this.props.deleteFavorite(item.id)
+                }
+              ],
+              { cancelable: false }
+            );
+          }
         }
-      ];  
+      ];
 
       return (
         <Swipeout right={rightButton} autoClose={true}>
